# reactmovies/src/data/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ReactMoviesAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/configuration/details": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbConfiguration"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbConfiguration"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbConfiguration"
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/toprated": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TopRatedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopRatedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopRatedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/newreleases": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewReleaseMovies"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewReleaseMovies"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewReleaseMovies"
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/nowplaying": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Dates": {
        "type": "object",
        "properties": {
          "maximum": {
            "type": "string",
            "nullable": true
          },
          "minimum": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NewReleaseMovies": {
        "type": "object",
        "properties": {
          "dates": {
            "$ref": "#/components/schemas/Dates"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewReleaseResult"
            },
            "nullable": true
          },
          "total_pages": {
            "type": "integer",
            "format": "int32"
          },
          "total_results": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "NewReleaseResult": {
        "type": "object",
        "properties": {
          "adult": {
            "type": "boolean"
          },
          "backdrop_path": {
            "type": "string",
            "nullable": true
          },
          "genre_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "original_language": {
            "type": "string",
            "nullable": true
          },
          "original_title": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "popularity": {
            "type": "number",
            "format": "double"
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "video": {
            "type": "boolean"
          },
          "vote_average": {
            "type": "number",
            "format": "double"
          },
          "vote_count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TmdbConfiguration": {
        "type": "object",
        "properties": {
          "images": {
            "$ref": "#/components/schemas/TmdbImages"
          },
          "change_keys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TmdbImages": {
        "type": "object",
        "properties": {
          "base_url": {
            "type": "string",
            "nullable": true
          },
          "secure_base_url": {
            "type": "string",
            "nullable": true
          },
          "backdrop_sizes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "logo_sizes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "poster_sizes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "profile_sizes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "still_sizes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TopRatedMovies": {
        "type": "object",
        "properties": {
          "adult": {
            "type": "boolean"
          },
          "backdrop_path": {
            "type": "string",
            "nullable": true
          },
          "genre_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "original_language": {
            "type": "string",
            "nullable": true
          },
          "original_title": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "popularity": {
            "type": "number",
            "format": "double"
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "video": {
            "type": "boolean"
          },
          "vote_average": {
            "type": "number",
            "format": "double"
          },
          "vote_count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TopRatedResult": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopRatedMovies"
            },
            "nullable": true
          },
          "total_pages": {
            "type": "integer",
            "format": "int32"
          },
          "total_results": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}